Fixed Windows automated build/install

+++ code/src/drivers/MySQL/MySQLStatement.cpp
@@ -263,7 +263,7 @@ void MySQLStatement::setParameterValues()
         case VAR_STRING:
         case VAR_TEXT:
         case VAR_BUFFER:
-            m_paramLengths[paramIndex] = param->dataSize();
+            m_paramLengths[paramIndex] = (unsigned long) param->dataSize();
             break;
 
         case VAR_DATE:
@@ -292,7 +292,7 @@ void MySQLStatement::setParameterValues()
 
 void MySQLStatement::MySQLStatement::prepare(const string& sql)
 {
-    if (mysql_stmt_prepare(statement(), sql.c_str(), sql.length()) != 0)
+    if (mysql_stmt_prepare(statement(), sql.c_str(), (unsigned long) sql.length()) != 0)
         throwMySQLError();
 }
 
@@ -548,7 +548,7 @@ void MySQLStatement::readPreparedResultRow(FieldList& fields)
                     bind.buffer_length = remainingBytes;
                     if (mysql_stmt_fetch_column(statement(), &bind, fieldIndex, offset) != 0)
                         throwMySQLError();
-                    bind.buffer_length = field->bufferSize();
+                    bind.buffer_length = (unsigned long) field->bufferSize();
                     bind.buffer = (void*) field->getBuffer();
                     fieldSizeChanged = true;
                 }
